refactor: src 구조 정리 (#5)
