+++ src/app/(main)/(start)/_components/NicknameStep.tsx
@@ -1,16 +1,16 @@
 'use client';
 
 import { useState } from 'react';
-import type { TrainerData } from '@/types/trainer';
+import type { TrainerData } from '../_types/trainer';
 import { motion } from 'framer-motion';
 import Image from 'next/image';
 import Typewriter from 'typewriter-effect';
-import { useLocalStorage } from '@/hooks/useLocalStorage';
-import { storageKeys } from '@/constants/key';
-import { nicknameSchema } from '@/schemas/nicknameSchema';
+import { useLocalStorage } from '@/shared/hooks/useLocalStorage';
+import { storageKeys } from '../_constants/key';
+import { nicknameSchema } from '../_schemas/nicknameSchema';
 import DialogBox from '@/shared/components/DialogBox';
 import TextField from '@/shared/components/TextField';
-import { landingMessages } from '@/constants/messages';
+import { landingMessages } from '../_constants/messages';
 
 export default function NicknameStep() {
   const [nickname, setNickname] = useState('');
