+++ src/app/(main)/(start)/_schemas/nicknameSchema.ts
