+++ src/app/(main)/(start)/_types/trainer.ts
