+++ src/app/(main)/(start)/page.tsx
@@ -1,4 +1,4 @@
-import NicknameStep from '@/components/start/NicknameStep';
+import NicknameStep from './_components/NicknameStep';
 
 export default function StartPage() {
   return <NicknameStep />;
