+++ src/shared/types/svg.d.ts
